+++ CHANGES.md
@@ -41,6 +41,7 @@ Features:
 
 Changes:
 - Abstract validation functions out of `AbstractSite.download_content` into reusable utils #1882
+- Skip `should_have_results` error during backscraping #1886
 
 Fixes:
 - Fix `lactapp_3` opinion download by using urllib instead of httpx to bypass Cloudflare TLS fingerprinting #1882